+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.5.9 - Unreleased
 
+### Changed
+* CocoImage and CocoAsset use the new AliasDictProxy interface to smoothly
+  transition between name conventions.
+* The "assets" key is now an alias of "auxiliary" for CocoImage and we will switch to "assets" as the primary in the future.
+
 
 ## Version 0.5.8 - Released 2023-03-17